Day 3: Basic Probability Puzzles #6
Objective
In this challenge, we practice calculating probability.

Task
Bag  contains  white balls and  black balls. Bag  contains  white balls and  black balls. You draw  ball from bag  and, without observing its color, put it into bag . Now, if a ball is drawn from bag , find the probability that it is black.

Output Format

In the editor below, submit your answer as Plain Text in the form of an irreducible fraction , where  and  are both integers.

Your answer should resemble something like:

3/4  
(This is NOT the answer, just a demonstration of the answer format.)