Standard Deviation Puzzles - 3
Let  and  be two independent "normal" random deviates with mean and standard deviation 10 and 3, and 20 and 4, respectively. Let  be the value of the standard deviation of the distribution obtained by adding the two distributions ( + ).

In the answer box, enter the value of , correct to one place of decimal.

Submission Modes and Output Format
You may either solve the problem on pen-and-paper and submit the final answer in the text box, or you may submit an R or Python program to accomplish the above task. Your output should be a floating point/decimal number, correct to 1 place of decimal.

In the text box below, enter a floating point/decimal number, correct to 1 place of decimal.

Alternatively, you may submit an R program, which uses the above parameters (hard-coded) and computes the answer.

Your answer should resemble something like:

9.1
(This is NOT the answer, just a demonstration of what the answering format should resemble).