Basic Probability Puzzles #10
Bill and Nina appear for an interview for two vacancies. The probability of Bill’s selection is 1/3 and that of Nina’s selection is 1/5. Find the probability that only one of them will be selected.

Fill in the answer in the text box, in the form of an irreducible fraction A/B where A and B are two integers.

e.g. Your answer should resemble something like:

3/4  
(This is NOT the answer, just a demonstration of what the answering format should resemble).