Day 1: Standard Deviation Puzzles #2
Objective
In this challenge, we practice calculating standard deviation.

Task
The heights of a group of children are measured. The resulting data has a mean of  meters, and a standard deviation of  meters. One particular child is  centimeters tall. Compute , the number of standard deviations away from the mean that the particular child is.

Enter the value of , correct to a scale of two decimal places.

Output Format

Your output must be a floating point/decimal number, correct to a scale of  decimal places. You can submit solutions in either of the  following ways:

Solve the problem manually and submit your result as Plain Text.

Submit an R or Python program, which uses the above parameters (hard-coded), and computes the answer.

Your answer format should resemble something like:

4.23
(This is NOT the answer, just a demonstration of the answering format.)